+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 Below you can find the changes included in each release.
 
+
+# 8.0.0
+Provide a system credential mechanism to authenticate users interacting with the DPR library. Endpoints can new be secured with a
+system token and caseloads and roles are checked my callback to the manage user api service rather than token inspection.
+
+# 7.13.0
+Migrate to call manage-users-api instead of the nomis-user-roles-api to retrieve caseloads.
+
+# 7.10.8
+New dashboard execution cancellation endpoint to support running dashboards for legacy nomis/bodmis reports.
+
+# 7.10.7
+New dashboard status endpoint to support running dashboards for legacy nomis/bodmis reports. 
+
 # 7.10.6
 Fixed issue with more than one multiselect filters in which the same filter value was applied to different columns.